move ComponentProvider to utils
