+++ quisp/utils/utils.h
@@ -1,0 +1,20 @@
+#ifndef QUISP_UTILS_H_
+#define QUISP_UTILS_H_
+
+#include "modules/QNIC.h"
+#include "modules/QNIC/StationaryQubit/StationaryQubit.h"
+#include "omnetpp.h"
+
+namespace quisp {
+namespace utils {
+
+using modules::QNIC_N;
+using modules::QNIC_names;
+using modules::QNIC_type;
+using modules::StationaryQubit;
+using omnetpp::cModule;
+
+}  // namespace utils
+}  // namespace quisp
+
+#endif
